implementing libxapi in fab/liblistwise

+++ fab/bake/bake.c
@@ -139,7 +139,7 @@ int bake_bp(
 	i = 0;
 	for(x = 0; x < bp->stages_l; x++)
 	{
-		log(L_BP | L_BPINFO, "STAGE %d of %d executes %d of %d", x, bp->stages_l - 1, bp->stages[x].evals_l, tot);
+		logf(L_BP | L_BPINFO, "STAGE %d of %d executes %d of %d", x, bp->stages_l - 1, bp->stages[x].evals_l, tot);
 
 		if(x)
 			dprintf(fd, "\n");
